refactor(rust): extract cr_patch helpers depuis cr_handlers (≤300 LOC)
